feat(commonware-node): serve consensus metrics (#212)
Spawns a server to serve requests on the metrics port provided in the consensus configuration. For the time being this needs to serve metrics on a port other then reth's because commonware does not integrate with the metrics family of crates.

Also makes the `metrics_port` setting optional: if no port is provided, no metrics will be served.

Pretty much taken straight from commonware [1], because their `commonware_runtime::tokio::telemetry::init` also tries to install a global tracing subscriber, which clashes with our current reth setup.

[1]: https://github.com/commonwarexyz/monorepo/blob/27ddd9e51ee6b8788cf39cbe78d248b58408006d/runtime/src/tokio/telemetry.rs#L87-L108